pyproject.toml
[tool.poetry]
name = "sales-late-day-tokens-service-deliver-order"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pymysql = "^1.1.0"
aio-pika = "^9.3.1"
asyncio = "^3.4.3"
asyncmy = "^0.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
